requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pypdf2>=3.0.0
httpx>=0.25.0
pydantic>=2.5.0
slowapi>=0.1.9
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
python-dotenv>=1.0.0
